# Refactored determinism tests using testhelpers