--- a/.gitignore
+++ b/.gitignore
@@ -2,6 +2,9 @@
 maps/
 *.swp
 
+# local replay file from tests/test_deterministic_replay.py
+tests/replay_local*.pickle
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]