--- a/tests/test_deterministic_replay.py
+++ b/tests/test_deterministic_replay.py
@@ -2,83 +2,95 @@ from pdb import set_trace as T
 import unittest
 from tqdm import tqdm
 
-import pickle
+import pickle, os, glob
 import random
 
 import nmmo
 
-# try to reuse functions defined in test_determinism.py
-import sys
-sys.path.append('tests')
+from testhelpers import TestEnv, TestConfig, serialize_actions, are_observations_equal
 
-from test_determinism import TestEnv, TestConfig, serialize_actions, are_observations_equal
+TEST_HORIZON = 50
+LOCAL_REPLAY = 'tests/replay_local.pickle'
 
+def load_replay_file(replay_file):
+   # load the pickle file
+   with open(replay_file, 'rb') as handle:
+      ref_data = pickle.load(handle)
+   
+      seed = ref_data['seed']
+      config = ref_data['config']
+      init_obs = ref_data['init_obs']
+      actions = ref_data['actions']
+      final_obs = ref_data['final_obs']
+      final_npcs = ref_data['final_npcs']
 
-def load_replay_file(replay_file='tests/deterministic_replay_ver_1.6.0.7_seed_5554.pickle'):
-   '''
-      This function will try to load the passed pickle file.
-      If the passed file is not found or not loaded properly, this function will generate a new file.
-      It is possible to supply a different file, but it should also be supported by the test class (TODO)
-   '''
-   try:
-      # load the pickle file
-      with open(replay_file, 'rb') as handle:
-         ref_data = pickle.load(handle)
-      
-         seed = ref_data['seed']
-         config = ref_data['config']
-         init_obs = ref_data['init_obs']
-         actions = ref_data['actions']
-         final_obs = ref_data['final_obs']
-         final_npcs = ref_data['final_npcs']
-
-         # test whether the loaded seed and config are valid
-         print('[TestDetReplay] Testing whether the seed and config are valid')
-         env_src = TestEnv(config, seed)
-         obs = env_src.reset()
-         assert are_observations_equal(obs, init_obs), "Something wrong with the provided pickle data"
-
-   except:
-      # generate the new data with a new env
-      seed = random.randint(0, 10000)
-      print('[TestDetReplay] Creating a new replay file with seed', seed)
-      config = TestConfig()
+      # test whether the loaded seed and config are valid
+      print('[TestDetReplay] Testing whether the seed and config are valid')
       env_src = TestEnv(config, seed)
-      init_obs = env_src.reset()
-
-      test_horizon = 50
-      actions = []
-      print('Running', test_horizon, 'tikcs')
-      for t in tqdm(range(test_horizon)):
-         actions.append(serialize_actions(env_src.realm, env_src.actions))
-         nxt_obs, _, _, _ = env_src.step({})
-      final_obs = nxt_obs
-      final_npcs = {}
-      for nid, npc in list(env_src.realm.npcs.items()):
-         final_npcs[nid] = npc.packet()
-         del final_npcs[nid]['alive'] # to use the same 'are_observations_equal' function
-
-      # save to the file
-      with open(replay_file, 'wb') as handle:
-         ref_data = {}
-         ref_data['version'] = nmmo.__version__ # just in case
-         ref_data['seed'] = seed
-         ref_data['config'] = config
-         ref_data['init_obs'] = init_obs
-         ref_data['actions'] = actions
-         ref_data['final_obs'] = final_obs
-         ref_data['final_npcs'] = final_npcs
-
-         pickle.dump(ref_data, handle)
+      obs = env_src.reset()
+      assert are_observations_equal(obs, init_obs), "Something wrong with the provided pickle data"
 
    return seed, config, actions, final_obs, final_npcs
+
+
+def generate_replay_file(replay_file, test_horizon):
+   # generate the new data with a new env
+   seed = random.randint(0, 10000)
+   print('[TestDetReplay] Creating a new replay file with seed', seed)
+   config = TestConfig()
+   env_src = TestEnv(config, seed)
+   init_obs = env_src.reset()
+
+   actions = []
+   print('Running', test_horizon, 'tikcs')
+   for t in tqdm(range(test_horizon)):
+      actions.append(serialize_actions(env_src, env_src.actions))
+      nxt_obs, _, _, _ = env_src.step({})
+   final_obs = nxt_obs
+   final_npcs = {}
+   for nid, npc in list(env_src.realm.npcs.items()):
+      final_npcs[nid] = npc.packet()
+      del final_npcs[nid]['alive'] # to use the same 'are_observations_equal' function
+
+   # save to the file
+   with open(replay_file, 'wb') as handle:
+      ref_data = {}
+      ref_data['version'] = nmmo.__version__ # just in case
+      ref_data['seed'] = seed
+      ref_data['config'] = config
+      ref_data['init_obs'] = init_obs
+      ref_data['actions'] = actions
+      ref_data['final_obs'] = final_obs
+      ref_data['final_npcs'] = final_npcs
+
+      pickle.dump(ref_data, handle)
+
+   return seed, config, actions, final_obs, final_npcs   
 
 
 class TestDeterministicReplay(unittest.TestCase):
    @classmethod
    def setUpClass(cls):
-      # TODO: allow providing the replay file by passing the file name to load_replay_file
-      cls.seed, cls.config, cls.actions, cls.final_obs_src, cls.final_npcs_src = load_replay_file()
+      """
+         First, check if there is a replay file on the repo, the name of which must start with 'replay_repo_'
+            If there is one, use it.
+
+         Second, check if there a local replay file, which should be named 'replay_local.pickle'
+            If there is one, use it. If not create one.
+
+         TODO: allow passing a different replay file
+      """
+      # first, look for the repo replay file
+      replay_files = glob.glob(os.path.join('tests', 'replay_repo_*.pickle'))
+      if replay_files:
+         # there may be several, but we only take the first one [0]
+         cls.seed, cls.config, cls.actions, cls.final_obs_src, cls.final_npcs_src = load_replay_file(replay_files[0])
+      else:
+         # if there is no repo replay file, then go with the default local file
+         try:
+            cls.seed, cls.config, cls.actions, cls.final_obs_src, cls.final_npcs_src = load_replay_file(LOCAL_REPLAY)
+         except:
+            cls.seed, cls.config, cls.actions, cls.final_obs_src, cls.final_npcs_src = generate_replay_file(LOCAL_REPLAY, TEST_HORIZON)
       cls.horizon = len(cls.actions)
 
       print('[TestDetReplay] Setting up the replication env with seed', cls.seed)