1. `test_determinism.py` and `test_deterministic_replay.py` now import common functions from testhelpers.py

2. `test_deterministic_replay.py` can use the replay file from the repo or from local. 
  * First, it looks for a replay file named like replay_repo_~~.pickle, which we don't have now. If we think the env is stable, then we can create one and put it on the repo. After making any changes, we can run `test_determnistic_replay.py` to see if the changes affected the game play at all.
  * Then, it looks for a local replay file named `replay_local.pickle`, which should not be pushed to repo (thus included in `.gitignore`) If there is, the test loads the seed, config, action, obs and compares those against the reproduced outputs.
  * When there is no `replay_local.pickle` (which would be the case after a fresh clone), then it produces a new local replay file and uses it afterward.

3. The obsolete replay file `deterministic_replay_ver_1.6.0.7_seed_5554.pickle` was deleted after confirming the game play was changed by the vecobs patch.

4. `PriorityQueue` and `Queue` in aStar, gatherBFS, forageDijkstra were replaced with `heapq` and python lists to speed up (by removing the thread lock overhead). The `test_deterministic_replay.py` confirmed that these changes did NOT affect the game play.
  * `aStar`: 8ms -> 6ms per call
  * `gatherBFS`: 10ms -> 8ms per call
  * `forageDijkstra`: 13ms -> 10ms per call