@kywch 2023-01-02
There are ongoing issues that may affect the game play. If these are solved/checked/acknowledged, then we may want to produce a deterministic replay file and push it to the repo.

* Item inventory error in the `test_api.py`: https://github.com/CarperAI/nmmo-environment/pull/7#issuecomment-1366184077
* The spawn functions in `realm.reset()`: in the vec-obs branch, realm.reset() has new self.players.spawn() and self.npcs.spawn(), which spawn npcs and let them move during the env.reset(). Please check.
* The order of entity/tile observations was changed in the vec-obs branch, and this affects targeting. For example, the closestTarget function in scripted/attack.py returns different results between original and vec-obs. I think we can leave it as is.