@kywch 2023-01-09
`env.reset()` was refactored to return the obs without doing the full `env.step()`. 

There are some cases where `env.render()` is called right after `env.reset()`, like the minimal tutorial example, an if statement was added to the render() to check whether `self.packet` was present.