@jsuarez5341 2023-01-09
This is a big help.

David is still working on the item bug. We're hoping to resolve that this week, lock in a stable version, and start on baselines.

Nicely done with the optimizations as well!